consolidate breadcrumbs, add little icons
